Slope limiter thinks it hits 3% of peak
Key optimisation is to enlarge the problem I think and to move the timer
inside the generated code. I think we are timing a fair bit of extra
stuff.

In theory we should be maybe a factor of 5 of peak when things like
streaming extra arrays and poor locality will hurt us.